+++ src/commonTest/kotlin/kotools/types/collection/NotEmptyListTest.kt
@@ -20,7 +20,7 @@ class NotEmptyListTest {
             .asNotEmptyList
             .getOrThrow()
         val result: Int = elements.head
-        result shouldEqual elements.asList.first()
+        result shouldEqual elements.toList().first()
     }
 
     @Test
@@ -30,7 +30,7 @@ class NotEmptyListTest {
             .getOrThrow()
         val result: NotEmptyList<Int>? = elements.tail
         result.shouldBeNotNull()
-            .asList contentShouldEqual elements.asList.drop(1)
+            .toList() contentShouldEqual elements.toList().drop(1)
     }
 
     @Test
@@ -40,19 +40,20 @@ class NotEmptyListTest {
     }
 
     @Test
-    fun asList_should_return_all_elements_as_a_List() {
-        val elements: List<Int> = List(3) { Random.nextInt() }
-        val result: List<Int> = elements.asNotEmptyList.getOrThrow().asList
-        result contentShouldEqual elements
-    }
-
-    @Test
     fun size_should_return_the_size_of_this_list_as_a_StrictlyPositiveInt() {
         val elements: NotEmptyList<Int> = List(3) { Random.nextInt() }
             .asNotEmptyList
             .getOrThrow()
         val result: StrictlyPositiveInt = elements.size
-        result.toInt() shouldEqual elements.asList.size
+        result.toInt() shouldEqual elements.toList().size
+    }
+
+    @Test
+    fun toList_should_return_all_elements_as_a_List() {
+        val elements: List<Int> = List(3) { Random.nextInt() }
+        val result: List<Int> = elements.asNotEmptyList.getOrThrow()
+            .toList()
+        result contentShouldEqual elements
     }
 
     @Test
@@ -60,14 +61,14 @@ class NotEmptyListTest {
         val elements: NotEmptyList<Int> = List(3) { Random.nextInt() }
             .asNotEmptyList
             .getOrThrow()
-        "$elements" shouldEqual "${elements.asList}"
+        "$elements" shouldEqual "${elements.toList()}"
     }
 
     @Test
     fun collection_asNotEmptyList_should_pass_with_a_not_empty_Collection() {
         val collection: Collection<Int> = List(3) { Random.nextInt() }
         val result: Result<NotEmptyList<Int>> = collection.asNotEmptyList
-        result.getOrThrow().asList contentShouldEqual collection
+        result.getOrThrow().toList() contentShouldEqual collection
     }
 
     @Test
@@ -84,7 +85,7 @@ class NotEmptyListTest {
         val tail: Array<Int> = List(2) { Random.nextInt() }
             .toTypedArray()
         val result: NotEmptyList<Int> = notEmptyListOf(head, *tail)
-        result.asList contentShouldEqual listOf(head) + tail
+        result.toList() contentShouldEqual listOf(head) + tail
     }
 }
 
@@ -104,7 +105,7 @@ class NotEmptyListSerializerTest {
             .asNotEmptyList
             .getOrThrow()
         val result: String = Json.encodeToString(elements)
-        result shouldEqual Json.encodeToString(elements.asList)
+        result shouldEqual Json.encodeToString(elements.toList())
     }
 
     @Test
@@ -112,7 +113,7 @@ class NotEmptyListSerializerTest {
         val collection: Collection<Int> = List(3) { Random.nextInt() }
         val encoded: String = Json.encodeToString(collection)
         val result: NotEmptyList<Int> = Json.decodeFromString(encoded)
-        result.asList contentShouldEqual collection
+        result.toList() contentShouldEqual collection
     }
 
     @Test